split Index
